Combine updatePlaylistStorage versions into one folder
