--- update-playlist-storage-v2/README.md
@@ -1,1 +1,0 @@
-Updates the public playlist JSON in Cloud Storage when NextUp adds a track to the playlist
