--- update-playlist-storage-v2/env.dev.yaml
@@ -1,3 +1,0 @@
----
-BUCKET: chirpradiodev-public
-FILE_NAME: playlist-test.json
