--- update-playlist-storage-v2/index.js
@@ -1,4 +1,0 @@
-const functions = require("@google-cloud/functions-framework");
-const updatePlaylistStorage = require("./lib/updatePlaylistStorage");
-
-functions.cloudEvent("updatePlaylistStorageV2", updatePlaylistStorage);
