--- update-playlist-storage-v2/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "update-playlist-storage-v2",
-  "version": "2.0.0",
-  "description": "Update playlist JSON feed in a public Cloud Storage bucket",
-  "main": "index.js",
-  "author": "Steve Lund <lund5000@gmail.com>",
-  "license": "MIT",
-  "scripts": {
-    "start": "yaml2json env.dev.yaml >.env.json && ./node_modules/.bin/env-cmd -f ./.env.json functions-framework --target=updatePlaylistStorageV2 --port=8082",
-    "deploy:dev": "gcloud functions deploy updatePlaylistStorageV2 --project chirpradiodev --gen2 --runtime=nodejs20 --region=us-central1 --env-vars-file env.dev.yaml --trigger-topic playlist-track-processed",
-    "test": "jest",
-    "lint": "npx eslint . --fix && npx prettier . --write"
-  },
-  "dependencies": {
-    "@google-cloud/functions-framework": "^3.3.0",
-    "@google-cloud/pubsub": "^4.3.2",
-    "@google-cloud/storage": "^7.7.0",
-    "luxon": "^3.4.4"
-  },
-  "devDependencies": {
-    "env-cmd": "^10.1.0",
-    "eslint": "^8.56.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-config-standard": "^17.1.0",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-n": "^16.6.2",
-    "eslint-plugin-promise": "^6.1.1",
-    "jest": "^27.5.1",
-    "prettier": "3.2.5",
-    "yaml2json": "^1.0.2"
-  }
-}
