+++ update-playlist-storage/.eslintrc.js
