+++ update-playlist-storage/.prettierrc
