+++ update-playlist-storage/__mocks__/@google-cloud/storage.js
