+++ update-playlist-storage/env.dev.yaml
@@ -1,2 +1,4 @@
-BUCKET: chirpradiodev-public
-FILE_NAME: playlist-test.json
+---
+  BUCKET: chirpradiodev-public
+  FILE_NAME: playlist-test.json
+  PLAYLIST_URL: https://chirpradiodev.appspot.com/api/current_playlist
