+++ update-playlist-storage/lib/formatTrack.js
