+++ update-playlist-storage/lib/loadJson.js
