+++ update-playlist-storage/lib/updateNowPlaying.js
