+++ update-playlist-storage/package.json
@@ -1,28 +1,34 @@
 {
-  "name": "update-playlist-storage",
-  "version": "1.1.0",
-  "description": "Retrieve and store current playlist payload in a public Cloud Storage bucket",
+  "name": "update-playlist-storage-v2",
+  "version": "2.0.0",
+  "description": "Update playlist JSON feed in a public Cloud Storage bucket",
   "main": "index.js",
   "author": "Steve Lund <lund5000@gmail.com>",
   "license": "MIT",
   "scripts": {
-    "start": "functions-framework --target=updatePlaylistStorage",
-    "deploy:dev": "gcloud functions deploy updatePlaylistStorage --project chirpradiodev --runtime=nodejs20 --trigger-topic update_playlist_storage --docker-registry=artifact-registry",
-    "deploy": "gcloud functions deploy updatePlaylistStorage --project chirpradio-hrd --runtime=nodejs20 --trigger-topic update_playlist_storage --docker-registry=artifact-registry"
+    "start:djdb": "yaml2json env.dev.yaml >.env.json && ./node_modules/.bin/env-cmd -f ./.env.json functions-framework --target=updatePlaylistStorage --port=8082",
+    "start:pubsub": "yaml2json env.dev.yaml >.env.json && ./node_modules/.bin/env-cmd -f ./.env.json functions-framework --target=updatePlaylistStorageV2 --port=8082",
+    "deploy:dev:djdb": "gcloud functions deploy updatePlaylistStorage --project chirpradiodev --runtime=nodejs20 --region=us-central1 --env-vars-file env.dev.yaml --trigger-topic update_playlist_storage",
+    "deploy:dev:pubsub": "gcloud functions deploy updatePlaylistStorageV2 --project chirpradiodev --gen2 --runtime=nodejs20 --region=us-central1 --env-vars-file env.dev.yaml --trigger-topic playlist-track-processed",
+    "test": "jest",
+    "lint": "npx eslint . --fix && npx prettier . --write"
   },
   "dependencies": {
-    "@google-cloud/storage": "^5.4.0",
-    "@google-cloud/functions-framework": "^3.3.0"
+    "@google-cloud/functions-framework": "^3.3.0",
+    "@google-cloud/pubsub": "^4.3.2",
+    "@google-cloud/storage": "^7.7.0",
+    "luxon": "^3.4.4"
   },
   "devDependencies": {
-    "eslint": "^7.0.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-config-standard": "^14.1.1",
-    "eslint-plugin-import": "^2.20.2",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prettier": "^3.1.3",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-standard": "^4.0.1",
-    "prettier": "^2.0.5"
+    "env-cmd": "^10.1.0",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-config-standard": "^17.1.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-n": "^16.6.2",
+    "eslint-plugin-promise": "^6.1.1",
+    "jest": "^27.5.1",
+    "prettier": "3.2.5",
+    "yaml2json": "^1.0.2"
   }
 }
