+++ update-playlist-storage/test/updateNowPlaying.spec.js
